+++ tests/Functional/UserControllerFunctionalTest.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace App\Tests\Controller;
+namespace App\Tests\Functional;
 
 use App\Repository\UserRepository;
-use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 use Exception;
+use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 
-class UserControllerTest extends WebTestCase
+class UserControllerFunctionalTest extends WebTestCase
 {
     const USER = 'user';
 
@@ -17,7 +17,6 @@ class UserControllerTest extends WebTestCase
     {
         $client = static::createClient();
 
-        // retrieve the test user
         $userRepository = static::getContainer()->get(UserRepository::class);
         $testUser = $userRepository->findOneBy(['username' => self::USER]);
         $client->loginUser($testUser);
@@ -26,19 +25,15 @@ class UserControllerTest extends WebTestCase
 
         $this->assertResponseIsSuccessful('Cannot find user edit page');
 
-        // Select the form and create a form object
         $form = $crawler->selectButton('Modifier')->form();
 
         $form['user_edit[password][first]'] = '123';
         $form['user_edit[password][second]'] = '123';
 
-        // Submit the form
         $client->submit($form);
 
-        // Follow the redirect
         $client->followRedirect();
 
-        // Check for a success flash message or other indicators of success on the redirected page
         $this->assertSelectorTextContains(
             'div.alert-success',
             'You successfully update your password',
