+++ tests/Smoke/DefaultControllerSmokeTest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Tests\Controller;
+namespace App\Tests\Smoke;
 
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 
-class DefaultControllerTest extends WebTestCase
+class DefaultControllerSmokeTest extends WebTestCase
 {
     public function testIndexRoute(): void
     {
