+++ tests/Unit/TaskTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Tests\Entity;
+namespace App\Tests\Unit;
 
 use App\Entity\Task;
 use App\Entity\User;
