+++ tests/Unit/UserTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Tests\Entity;
+namespace App\Tests\Unit;
 
 use App\Entity\User;
 use PHPUnit\Framework\TestCase;
